+++ philo_bonus/src/utils_bonus.c
@@ -6,7 +6,7 @@
 /*   By: tat-nguy <tat-nguy@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/01 22:20:19 by tat-nguy          #+#    #+#             */
-/*   Updated: 2025/05/31 23:46:54 by tat-nguy         ###   ########.fr       */
+/*   Updated: 2025/05/31 23:58:19 by tat-nguy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -67,38 +67,4 @@ void	print_error(char *message)
 {
 	printf("%s\n", message);
 	exit(EXIT_FAILURE);
-}
-
-void	update_last_meal(t_philo *philo)
-{
-	sem_wait(philo->meal_lock);
-	philo->last_meal = get_time_ms();
-	sem_post(philo->meal_lock);
-}
-
-void	increment_meals(t_philo *philo)
-{
-	sem_wait(philo->meal_lock);
-	philo->meals_eaten++;
-	sem_post(philo->meal_lock);
-}
-
-size_t	get_last_meal(t_philo *philo)
-{
-	size_t	time;
-
-	sem_wait(philo->meal_lock);
-	time = philo->last_meal;
-	sem_post(philo->meal_lock);
-	return time;
-}
-
-int	get_meals_eaten(t_philo *philo)
-{
-	int	meals;
-
-	sem_wait(philo->meal_lock);
-	meals = philo->meals_eaten;
-	sem_post(philo->meal_lock);
-	return meals;
 }
